comments

+++ src/app/component/cards/cards.component.ts
@@ -22,11 +22,13 @@ export class CardsComponent implements OnInit {
 
   @Input()
   artWork: ArtWork;
+  //default Artwork Image
   ArtWorkImageLink = 'https://images.bluethumb.com.au/uploads/listing/201464/debra-ryan-chakra-feather-modern-inspirational-artwork-bluethumb-01c5.jpg?w=766&h=766&fit=crop&auto=compress&cs=tinysrgb&q=70&s=9e0a3ef9469be01f1bc75e1e615d5caa';
 
   ngOnInit(): void {
     const t = new Date(1970, 0, 1);
-    // @ts-ignore
+    //converting seconds to a real date
+    // @ts-ignore used to ignore a ts remark
     t.setSeconds(this.artWork.Date.seconds);
     this.artWork.Date = t;
     this.fetchImage().then();
@@ -38,7 +40,7 @@ export class CardsComponent implements OnInit {
 
   async onDelete() {
     if (confirm('Are you sure to delete this artwork ?')) {
-      // const artWorkId=this.artWork.id;
+      // each one of them are being deleted one after others
       await this.symbolService.deleteSymbol(this.artWork.id);
       await this.ArtWorkService.deleteArtWorks(this.artWork);
       await this.directionService.deleteDirection(this.artWork.id);
@@ -51,30 +53,26 @@ export class CardsComponent implements OnInit {
   }
 
   handleOkMiddle(): void {
+    //getting values from card
     const Description = this.addArtWorkCardComponent.DescriptionInput.value;
     const Title = this.addArtWorkCardComponent.NameInput.value;
     const Author = this.addArtWorkCardComponent.AuthorInput.value;
 
     const image = this.addArtWorkCardComponent.UploaderInput.image;
-    console.log(image)
     const Urimage = (image?.name)?image.name:this.artWork.Urimage;
-    console.log(Urimage)
     const artWork = {Author, Date: new Date(this.addArtWorkCardComponent.DateInput.value), Description, Title, Urimage};
-    console.log(artWork);
 
-    if (image){
-      console.log("oui")
+    if (image){ //if the image has changed, you can upload a new one and then delete the old one
       this.imageStorageService.uploadImage('ArtImages', image.name, image).then(res =>
           this.ArtWorkService.updateArtWorks(this.artWork.id, artWork)
         );
       this.imageStorageService.deletePreviousImage('ArtImages', this.artWork.Urimage).then();
     }
-    else{
-      console.log("non")
+    else{ // Otherwise just update the artwork with some of the news values
       this.ArtWorkService.updateArtWorks(this.artWork.id, artWork);
     }
 
-
+  //modal visibility
     this.isVisibleMiddle = false;
   }
